+++ QBCore_2F0666.base/resources/[standalone]/rp-ignition/fxmanifest.lua
@@ -13,5 +13,4 @@ client_scripts {
 
 dependency 'qb-core'
 dependency 'qb-vehiclekeys'
-dependency 'qb-phone'
 dependency 'rp-chat'